Add logic for missing prefs
This means that iCloud is totally disabled, thus Find my Mac isn't on.

+++ Find My Mac Status.sh
@@ -5,6 +5,10 @@
 loggedInUser=`python -c 'from SystemConfiguration import SCDynamicStoreCopyConsoleUser; import sys; username = (SCDynamicStoreCopyConsoleUser(None, None, None) or [None])[0]; username = [username,""][username in [u"loginwindow", None, u""]]; sys.stdout.write(username + "\n");'`
 homedir=$(dscl . -read /Users/$loggedInUser NFSHomeDirectory | cut -d' ' -f2)
 
-FindMyMac=$(/usr/libexec/PlistBuddy -c "print :Accounts:0:Services" "$homedir"/Library/Preferences/MobileMeAccounts.plist | grep -B 1 FIND_MY_MAC | grep Enabled | awk '{print $3}')
+if [ ! -f "$homedir"/Library/Preferences/MobileMeAccounts.plist ]; then
+	echo "<result>false</result>"
+	exit
+fi
 
+FindMyMac=$(/usr/libexec/PlistBuddy -c "print :Accounts:0:Services" "$homedir"/Library/Preferences/MobileMeAccounts.plist | grep -B 1 FIND_MY_MAC | grep Enabled | awk '{print $3}')
 echo "<result>$FindMyMac</result>"
